config: arm64: enable CONFIG_IKCONFIG_PROC again
During the migration of the kernel from v5.15 -> v6.6 CONFIG_IKCONFIG_PROC
was unset in the arm64 config only. This breaks the arm64 support for the
usbipd-win project as it uses /proc/config.gz to obtain kernel configuration.

Please see: https://github.com/dorssel/usbipd-win/issues/1175

This patch enables CONFIG_IKCONFIG_PROC as was done in the v5.15 WSL arm64
kernel config:

https://github.com/microsoft/WSL2-Linux-Kernel/blob/linux-msft-wsl-5.15.y/arch/arm64/configs/config-wsl-arm64#L137

Which is currently unset in the v6.6 WSL arm64 kernel config:

https://github.com/microsoft/WSL2-Linux-Kernel/blob/linux-msft-wsl-6.6.y/arch/arm64/configs/config-wsl-arm64#L149

However, x86 config enables it in the v6.6 config:

https://github.com/microsoft/WSL2-Linux-Kernel/blob/linux-msft-wsl-6.6.y/arch/x86/configs/config-wsl#L168

Signed-off-by: Tyler Baker <tyler.baker@oss.qualcomm.com>